% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{igraph_to_networkD3}
\alias{igraph_to_networkD3}
\title{Convert igraph to JSON}
\source{
A direct import of \code{igraph_to_networkD3} from JJ Allaire, Peter
 Ellis, Christopher Gandrud, Kevin Kuo, B.W. Lewis, Jonathon Owen, Kenton
 Russell, Jennifer Rogers, Charles Sese, CJ Yetman (2015). \link{networkD3}:
 D3 JavaScript Network Graphs from R
}
\usage{
igraph_to_networkD3(edges, nodes, group, what = "both")
}
\arguments{
\item{...}{igraph object}
}
\description{
Convert igraph to JSON
}
