% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threeNet.R
\name{threeNet}
\alias{threeNet}
\title{Create 3D Network}
\usage{
threeNet(data, width = NULL, height = NULL, elementId = NULL,
  controlType = "trackball", nodeColors = "#f7a440", nodeSizes = 4,
  nodeOpacity = 0.75, nodeResolution = 8, linkLabel = "",
  linkVisibility = TRUE, linkColor = "#ffffff", linkWidth = 0)
}
\description{
This is the main function for the package which takes a list of network data
(see example) and creates a 3D network in HTMl and JS, using "3d-force-graph"
}
\examples{
bipartiteEdges <- data.frame(source = c("A","A","A", "B", "B","C", "B"),
                             target = c("B", "Oranges", "Apples", "Kiwis", "Bananas", "A", "A" ),
                             weight = c(10, 10, 10, 10, 10, 10, 10))

bipartiteNodes <- data.frame(id = unique(c(as.vector(bipartiteEdges$source),
                                           as.vector(bipartiteEdges$target))),
                            colors = rainbow(7))


bipartite <- threeNet::igraph_to_networkD3(bipartiteEdges, bipartiteNodes)
# Format of input data:
bipartite
threeNet::threeNet(jsonlite::toJSON(bipartite))
}
